refactor(accounts): extract AccountRegistryLoader load-orchestration collaborator (Wave 3 / 3a-4)
Fourth and LARGEST 3a collaborator split: the catalog LOAD orchestration — the
loadCatalogs stale-while-revalidate pipeline, the first-page/paginate network crawl +
direct-GET fallbacks, CP-D1 preload + slim-snapshot hydrate/carve/write, the owned
background-crawl task registry + its test-boundary drain choreography, and the
loading-completion handler dedupe — moves out of AccountsManager into an injected
AccountRegistryLoader. It ORCHESTRATES the already-extracted registryCache (3a-1) /
registryStore (3a-2) / authDocLoader (3a-3). Behaviour-identical. `final class`, not
actor (loadCatalogs/cancelAndDrainBackgroundWork called synchronously). Off develop
(all three prior extractions merged).

**AccountsManager 1910 → 955 LOC (−955).** New AccountRegistryLoader.swift (873 LOC).
Across the four extractions the hub is now 2383 → 955.

- Hub keeps thin facades (loadCatalogs, loadAccountSetsAndAuthDoc, preloadAccountsFromDiskCacheSync,
  the drain/join _*ForTesting seams, a settable snapshotResourceResolver forwarding property)
  + static shims (carveSlimFeed, buildAccountIndex) so AppContainer / CatalogCrawlScheduler /
  PalaceTestSetup / EVERY existing test are byte-identical — ZERO external edits (the
  single-module /rigorous-fix proof; confirmed by git status = 6 files only).
- **Drain-flag set-point (architect finding 2):** the hub cancelBackgroundWork() /
  cancelAndDrainBackgroundWork(timeout:) facades set `_explicitCancelCalled = true` FIRST,
  then delegate; the loader's cancel/drain bodies do NOT set the flag and do NOT call back
  for it. `_explicitCancelCalled` + the live-instance registry STAY on the hub, preserving
  the byte-identical 3a-3 AuthDocumentLoader.isTornDown binding. A verbatim relocation here
  would have reintroduced the exact test-pollution regression this wave prevents.
- 9 injected provider closures ([weak self]) incl. currentLibraryAccountProvider for the
  COPPA age-check (finding 3). lazy var (closures capture self — the authDocLoader precedent).

Verified: god-class freeze re-baselined DOWN 1910 → 955; locator 304 → 302; shared-read 213;
Palace-noDRM BUILD SUCCEEDED (production compile). Architect pre-review BLOCKED then APPROVED
after 3 contract corrections (preload/snapshotResolver facades, hub-sets-drain-flag, 9th closure).

**Scope:** the catalog load orchestration + owned-crawl + drain out of AccountsManager into
an injected collaborator; Accounts area only, single-module.
**Not done:** DRM PalaceTests + the FULL suite (the pollution-sensitive Cancellation /
FirstRunDecode / CacheRead / LaunchSnapshot / StateMachineWiring pins + mutation gate) run in
CI ONLY — worktree can't build the Adobe host. `currentUserAccountProvider`'s UNREACHABLE
[weak self]==nil fallback uses TPPUserAccount.sharedAccount() (no non-hub TPPUserAccount exists;
a strong capture would retain-cycle) — flagged for review.
**Deferred:** the last two 3a collaborators (CredentialResolver, CurrentAccountStore).
